BlameCommand: Notice when the file is unsaved:
Querying blame information for a file with unsaved modifications shouldn't be allowed. This is because the line-number of the text caret in the in-memory representation of the file has a high chance of referring to an unrelated line in the on-disk representation of the file. Since `git blame` operates solely with on-disk representation, bogus information could be shown in the phantom because of this discrepancy.

BlameShowAllCommand already protects against this.